fix: rearrange files, remove drift dep as we cannot compile it
