+++ examples/solver/drift-intent-solver/Cargo.toml
@@ -8,8 +8,9 @@ cosmwasm-std = "2.0.1"
 serde = { version = "1.0.145", default-features = false, features = ["derive"] }
 cosmwasm-schema = "2.0.1"
 ed25519-dalek = "1.0"
-drift = { git = "https://github.com/drift-labs/protocol-v2" }
 bs58 = "0.4"
+curve25519-dalek = "4.1.3"
+sha2 = "0.10.8"
 
 [lib]
 crate-type = ["cdylib", "rlib"]
