--- examples/solver/drift-solver/Cargo.toml
@@ -1,29 +1,0 @@
-[package]
-name = "amm-solver"
-version = "0.0.1"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-cosmwasm-std = "2.0.1"
-serde = { version = "1.0.145", default-features = false, features = ["derive"] }
-schemars = "0.8.10"
-cosmwasm-schema = "2.0.1"
-hex = "0.4.3"
-bs58 = "0.5.1"
-bech32 = "0.11.0"
-sha2 = "0.10.8"
-curve25519-dalek = "4.1.3"
-tiny-keccak = { version = "2.0", features = ["keccak"] }
-
-[patch.crates-io.curve25519-dalek]
-git = "https://github.com/solana-labs/curve25519-dalek.git"
-rev = "b500cdc2a920cd5bff9e2dd974d7b97349d61464"
-
-[lib]
-crate-type = ["cdylib", "rlib"]
-
-[features]
-# use library feature to disable all instantiate/execute/query exports
-library = []
