--- examples/solver/drift-solver/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod svm;
-pub mod astromesh;
